Hi all, I'm doing a change in the anuga structure, moving the code to  \anuga_core\source\anuga
After you have done an svn update, the PYTHONPATH has to be changed to;
PYTHONPATH = anuga_core/source/

This is part of changes required to make installation of anuga quicker and reducing the size of our sandpits.

If any imports are broken, try fixing them.  With adding anuga. to them for example.  If this seems to have really broken things, email/phone me. 


Cheers
Duncan