Brought back caching at the file_function level which was moved down in changeset:4278
I used the command svn merge -r4278:4277 util.py

The problem was that the input arguments to cahing were so large that it took a long time to hash them. File_function is much faster now and I believe the original problem was fixed when None, True and False were treated explicitly in cahing.myhash (changeset:4198 and changeset:4270). 

If a problem arises again, let us write a unit test and then deal with it.