Optimised gravity term making update_conserved_quantities faster. Instead of computing the average depth from vertices we now use the centroid values directly.
On the Linux machine nautilus the okushiri profile increased the speed of update_conserved_quantities from 4.52s to 2.8s or a 38% increase. The total time was about 38s making this optimisation a 5% overall improvement.

An experimental optimisation for flat beds did not yield any measurable 
improvements, so it was commented out.