Fixed unit test that broke between changeset:5846 and changeset:5847
I have convinced myself that the problem was that although the test passed 
it was wrong.
In changeset:5846 it is clear that the time boundary in the helper function 
create_sww_boundary was not evaluated correctly as domain.time got updated
after the call to update_boundary in domain.py.
This caused the boundary evaluation to lag.

However, this lag was not present originally in ANUGA so this error 
would have been introduced
somewhere along with the euler stepping - most likely around November 2007.

I added an assert checking that the lag is not present now or in the future.