Thought a bit more about the flux optimisation implemented in changeset:7105
The y-momentum limiter has been reinstated. Instead, I implemented a cut-off for limiting the momentum as described in the manual (see Section about flux limiter). For the simple profile compute_fluxes now takes 2.964s (as opposed to 4.284s prior to changeset:7105) and the overall runtime of evolve takes 9.797s (as opposed to 10.909s). For the okushiri example the timings are now 109.551s for compute_fluxes and 293.314s overall.

Overall this amounts to a 10% overall performance improvement.

See also changeset:6703 for past timings

These timings were conducted on a 3.0GHz Intel Core-2 running Ubuntu Linux.


I had to updated a few tests that used to compare to exact values that are now slightly different.