Have been playing with using a slope limited velocity to calculate fluxes (hence the addition of evolved_quantities as well as conserved quantities.
But the commit is to fix a problem Rudy found with sww2dem. Seems 
numpy.array2string  is a little too clever, in that it summarizes
output if there is a long sequence of zeros to 
[0.0, 0.0, 0.0, ... 0.0, 0.0 ,0.0] To get around this I have added 
a call to numpy.set_options(threshold=sys.max_int) to turn this 
behaviour off!